pyproject.toml
[tool.black]
exclude = '''
    (
        \.eggs
        | \.git
        | \.hg
        | \.mypy_cache
        | \.tox
        | \.venv
        | _build
        | buck-out
        | build
        | dist
        | reports
    )
    '''


[tool.poetry]
name = "dotcfg"
version = "0.1.0"
description = ""
authors = ["Alex Cano <alexjcano1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-box = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
mypy = "^0.790"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
